graphicsProvider: replace some functions with those provided by ComputerNerd

+++ src/graphicsProvider.hpp
@@ -15,7 +15,7 @@
 #include <math.h>
 
 int PrintMiniFix( int x, int y, const char*Msg, const int flags, const short color, const short bcolor ); 
-void plot(int x0, int y0, int color);
+void plot(int x0, int y0,unsigned short color);
 void drawRectangle(int x, int y, int width, int height, unsigned short color);
 void drawLine(int x1, int y1, int x2, int y2, int color);
 void VRAMReplaceColorInRect(int x, int y, int width, int height, color_t color_old, color_t color_new);
